+++ database/migrations/2026_02_25_134734_create_artisan_verification_codes_table.php
@@ -10,21 +10,23 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('artisan_verification_codes', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('task_id');
-            $table->foreign('task_id')->references('id')->on('artisan_tasks')->onDelete('cascade');
-            $table->string('code')->unique();
-            $table->unsignedBigInteger('landlord_id');
-            $table->unsignedBigInteger('artisan_id')->nullable();
-            $table->unsignedBigInteger('tenant_id')->nullable();
-            $table->timestamp('expires_at')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('artisan_verification_codes')) {
+            Schema::create('artisan_verification_codes', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('task_id');
+                $table->foreign('task_id')->references('id')->on('artisan_tasks')->onDelete('cascade');
+                $table->string('code')->unique();
+                $table->unsignedBigInteger('landlord_id');
+                $table->unsignedBigInteger('artisan_id')->nullable();
+                $table->unsignedBigInteger('tenant_id')->nullable();
+                $table->timestamp('expires_at')->nullable();
+                $table->timestamps();
 
-            $table->foreign('landlord_id')->references('user_id')->on('users')->onDelete('cascade');
-            $table->foreign('artisan_id')->references('user_id')->on('users')->onDelete('set null');
-            $table->foreign('tenant_id')->references('user_id')->on('users')->onDelete('set null');
-        });
+                $table->foreign('landlord_id')->references('user_id')->on('users')->onDelete('cascade');
+                $table->foreign('artisan_id')->references('user_id')->on('users')->onDelete('set null');
+                $table->foreign('tenant_id')->references('user_id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     /**
